Add simple install cache to spack applications
Currently each expierment tries to install spack packages and regenerate
modules. This is fine, but we can guarntee this work will already be
done, and thus we spend a lot of time in needless operations.

This PR is a first attempt at resolving that, and speeds things up
significantly. In my workspace with ~20 experiments it takes install
time from ~3mins to ~20s  (saves ~6-7s per repeated experiment)