Rework epander / experiment / application logic
This commit represents a major reworking of the internal structures for
housing experiments.

The basic goal of this commit is to create an experiment set that houses
all experiments within a workspace. This results in all of the
experiments existing in memory at the same time.

This change required changing the fundamentals of how the expander class
worked, as it no longer stores variables. Variables are now stored
within the application class (which also contains a reference to the
expander class).

This commit also updates the tests, and the `workspace info` command
based on the changes from how the experiments are stored.