Update the way spack_runner installs compilers
This commit changes how spack_runners install compilers. Now, the runner
checks the output of `spack compiler info <compiler>` before attempting
to install it.

One negative side effect of this is that compilers that are needed
within an environment have to be installed as independent packages.

However, this is the same as how spack currently handles it, so likely
this is preferred.

A test to ensure previously installed compilers are not installed fresh
again is added as well.