Fix color in workspace info, and prepare for cross experiment references
This commit fixes an issue in workspace info where variables that
contained the @ symbol would throw and error due to color.cprint
expecting a color identifer.

Additionally, the method is restructure to prepare for cross experiment
referneces. To aid this, all experiments are generated in the first
pass, and second (just to make printing nicer) the experiments are
generated again but only to scope variables properly.