Re-scope pytest import, and add exception handling
This commit moves the import of pytest to happen immediately before
pytest is used. This prevents an issue where the whole unit_test.py
module cannot be imported if pytest is missing, and allows Ramble to be
used without the pytest dependency.

Additionally, error messages are added to communicate that
requirements.txt should be installed if pytest is imported but not
found.