Add known keyword validation logic
This commit adds a new class (keywords) to house known ramble variable
names. Logic is added using this to perform validation on different sets
of variable definitions to ensure they are properly defined.

Tests are also added to make sure the proper error messages are passed
when varible names are incorrectly used.