Add define_package_paths phase
This commit moves package path definition from the install_software
phase to its own phase called define_package_paths.

Software installation is guarded by a cache (to help make the process
faster for repeated installations of the same environment). This
prevented package paths from being define in experiments in a series
(except for the first experiment in the series.

Fixes #95