Add fallback options to flake8 file list logic
This commit adds several fallback options to the flake8 file list logic.

The default is to use the provided base branch and flags. If that
doesn't work, `origin/develop` is used as the base branch with the other
flags set to their input values. Following that, `origin/main` is used.
And as a final step, all files are linted.