Generalize repository classes
This commit generalizes the repository classes in preparation for
holding other types (such as mods).

Primarily it converts the classes to use agnostic language
(application -> object) to describe what the repositories house, but it
also introduces ObjectTypes and type_definitions at the top which house
the specific types that are supported.