Add modifier type to repository
This commit updates the repository methods to include a modifier type in
addition to the default object and application types.

Additionally, it generalizes some of the type specific methods, updates
references, adds a schema for a modifier_repos file, and updates tests
to have modifier and application specific mock functionality.