Fix issue where we used ramble package name not spacks
Some applications need to know the path to the spack install directory
of the package. This appears as `{package_name}` in the
application.py `(eg {wrf}`). A recent change accidentally made it so
that the path was stored internally as a mapping to the
ramble_package_name (eg wrfv3) instead of the name by which the package
is known to spack (eg wrf).

This means are written wrfv3 and wrfv4 are broken on head, and this
addresses that.