Add support for two new success criteria methods
This commit adds the ability for applications to define an
`evaluate_success` function, which can perform application specific
validation of an experiment on top of the built in success criteria.

This function is automatically added to all experiments if it is
defined.

Additionally, success criteria can now define two new modes:
- mode: 'application_function' - Will evaluate with the
`evaluate_success` method from an application.py. This is not necessary,
as it is added by default anyway.
- mode: 'fom_comparison' - Will evaluate success with comparison logic
that can use variable definitions and refer to the FOM values.