Add support for prefix on spack commands
This commit adds support for a config:

  config:spack:install:prefix

Which can define a prefix to all `spack install` executions (allowing
parallel installs).

Additionally, this updates the format of the config config_section,
which can be applied using `ramble config update config`. This change is
also applied to the `concretize` command.

Tests and documentation are updated as well.