Add support for filtering with where statements
This commit performs a few important actions. The first and most
important is that it adds support for both `--where` and
`--exclude-where` flags on workspipeline commands
(such as `ramble worksapce setup`) which can be used to control which
experiments an action will be performed against.

The second and more internal thing this commit does is it reworks how
pipelines are defined. This moves pipelines from being a portion of a
workspace to being their own class which acts on a workspace instead.