Add directives for defining phases and their order
This commit adds a new directive:
`register_phase`

Which will create a new phase within a pipeline. This directive contains
an attribute named `depends_on` to define the phases a new phase defines
on to control the order more explicitly.

Additionally, the `--include-phase-dependencies` flag is added to all
workspace pipeline commands, to allow Ramble to automatically include
dependencies of phases when the phases to execute are filtered.

Tests are added as well.