Allow empty spack environments when constructing hash
When building an experiment / workspace inventory, previously it was
required that if an environment had a spack.yaml, it also created a
spack.lock (or Ramble would error).

This commit allows spack.yaml files that have no packages (And thus,
won't concretized into a spack.lock) without throwing an error.