Fix expander removing leading zeros
This commit fixes an issue where the expander would remove leading
zeros when passed a floating point number (as a string).

These can be forced to remove leading zeros still by using curly braces
and a format string, however there was no way to force leading zeros to
not be removed previously.