Add logic to build no_expand_vars for applications
This commit allows workload_variables to define `expandable=False` which
prevents expansion of a variable definition, when it is used. This also
disables math evaluation, and explicit formatting of the values.

Additionally, applications now build a set of no_expand_vars and pass
this to the expander to prevent expansion of variables defined in this
fashion.