Add support for defining tags on workloads, and experiments
This commit adds the ability for workload directives to define tags,
which are then tracked in experiment instances.

Additionally, a `tags` attribute is added to the YAML schema for
experiments and workloads. This allows experiments to be tagged at a
finter grain within a workspace.

The workspace info command is also updated to provide more control over
the information provided.