Change spack hash semantics to hash yaml contents
This commit changes the semantics of how hashes of spack environments
work. Previously, spack.lock files were hashed when an experiment was
setup, however this gets in the way of many things (like partially
setting up a workspace).

This commit changes the hashing to only hash the contents of the YAML
file (not actually the YAML file, as we don't always write it) instead
to avoid these issues.