Add a deployment command and push pipeline
This commit adds a deployment command with push and pull as
subcommands, and a push deployment pipeline.

The deployment command is useful to push a workspace into a fully
contained deployment to manage experiments in a remote location.

The push pipeline is used to create a deployment from a
ramble workspace. This encapsulates all of the artifacts the workspace
needed to perform its experiments. It captures the workspace's
configuration files, and creates a workspace specific object repository
that contains package, application, and modifier definitions that the
workspace needs to execute.

The push subcommand can create a tar of the deployment, and also push
the deployment to a remote location.

The fetch subcommand takes a path to a remote deployment, and overwrites
the contents of the current workspace with the deployment's.