Update Variables schema to an object not an array
The current schema implies variables is a list (dashes), which is not
true. This change fixes that and instead updates it to an object.

The list was no causing functional issues at run time, but did manifest
as when using ramble to update configs, as it was adding elemnts in a
list, eg:

```
ramble -c 'variables:batch_submit:""' config blame variables
```

Was returning:

```
batch_submit:
- ''
```

This change fixes that, and instead it now returns

```
batch_submit: ''
```