Convert environment package name rendering to lazy
This commit converts the rendering of package names within an
environment to lazy. Previously, template packages were search for when
building template environments. This caused issues if a variable package
name was passed in (and expected to render later at experiment time).

This commit changes this behavior to only error when the environment is
actually rendered rather than when the templates are defined.