Fix removal of scalars, and ignoring modifier variables
This commit fixes two issues with removal of unused variables, primarily
related to removing variables that modifiers might use.

The first fix updates the removal code to only remove list variables (as
they are problematic for rendering experiments).

The second fix propagates variables defined in a modifier (using the
modifier_variable directive) as used variables even if they are not used
in an expansion to ensure they are not removed accidentally.