Allow expander to return string representation of invalid operators
This commit updates the way binary operators work in the Expander.
Previously, if a string with binary operators was passed in (i.e.
`intel-mpi-benchmarks`) the binary operators were expected to actually
operate on numerical inputs. This caused unescaped string comparisons
(i.e. `{application_name} == intel-mpi-benchmarks`) to fail as the
expander would give a syntax error and not evaluate this statement.

This commit changes this behaivor to force the expander to return a
string representation of the args and binary operator:
(i.e. `intel-mpi-benchmarks` is returned as `'intel-mpi-benchmarks'`) to
allow these types of comparisons.