Simplify user experience for managing repositories
This commit updates the main user experience for interacting with object
repositories (applications and modifiers).

Previously, each repository required a configuration file for each
object type it held (repo.yaml for applications, modifier_repo.yaml for
modifiers). This commit updates this to allow the `repo.yaml` file to be
read for both applications and modifiers. Optionally, applications can
be configured with an application_repo.yaml file, while modifiers can be
configured with modifier_repo.yaml.

The `ramble repo` command is also updated to allow `-t any` as a valid
type, in which case ramble will manipulate all object types that appear
valid for a repository.