Update variable usage tracking to support passthrough lists
This commit updates the logic around tracking variables based on if they
are rendered or not (and selectively removing variable definitions).

The new logic allows variables that are not rendered as part of the
typical make_experiments phase (i.e. part of keyword variables, or
template definitions) do not cause an error when they have length
conflicts with other lists (and are left as explicit lists in their
value definitions).