Add a `ramble style` command
This heavily models after `spack style`, and includes only `flake8`
check currently. The intent is to have a general style check command to allow for
adding in more style check tools like `black`.

`ramble style` is a replacement of `ramble flake8` as of now.

Next step is to add in black formatter support.