Allow modifiers to influence experiment variables
This adds a hook in modifiers called modify_eperiment, which can be used
to change experiment variables before the experiment is completely
setup.

This allows modifiers to change the value of important variables (such
as n_ranks) and have that change reflected in things like the name of
the experiment.