Fix environment creation for spack package manager
This commit updates the environment creation logic for the spack package
managers. Previously, environment creation only happened when not
performing setup as a dry-run. Additionally, installation only happened
if an environment was required (which meant the application had software
specs defined in the application definition).

This commit fixes these issues.