Expand logs before adding to the purge set
This commit updates the set building logic for determining which logs
need to be purged at the beginning of an experiment. Previously, the raw
strings were used (and later expaned as part of a template) however
duplicate entries were then added.

This commit expands log files that are added to the purge set to ensure
duplicate entries are not added.