Test activate shortcut in the tutorial test
Currently the implementation of the activate shortcut (`-a`) for `workspace
create` is fragile: it relies on the stdout from the create command, so
any redundant log entries would break it. There's no current end-to-end
test to catch its potential breakage (challenging to test given the
impl lives in the shell wrapper layer.) So piggy-back on this tutorial
e2e test to ensure its working.