Add support for arbitrary commands in formatted executables
This commit allows formatted_executable definitions to have a `commands`
attribute, which can define the lines that should be in the formatted
output.

The default is to format all of the commands for the experiment, but
this can be override to format arbitrary strings / variables. Item in
the `commands` list will be expanded as a variable definition, and then
split over new-lines before formatting based on the formatted executable
definition.