Update spack and ramble install modifiers
This commit updates the install-spack and install-ramble modifiers to
have quite execution modes which can be used to suppress injecting
commands into the resulting `command` variable. This is useful when
creating formatted executables with the definitions provided by these
modifiers.

Additionally, some of the formatting of the explicit code blocks are
updated.