Update repos used in deployments
This commit updates how repositories are used within deployments.

Previously, when creating a deployment, Ramble would create
configuration files to add the deployment into the resulting workspace.

This commit changes this behavior so that the repository is created, but
it is added at deployment pull time instead of deployment push time.
This allows the path of the deployment respository to be absolute on the
remote machine rather than a relative path that hopefully will work on
the remote machine.