Allow include paths to refer to workspace paths
This commit adds a workspace dictionary named `workspace_paths` which
has common replacement paths for getting absolute paths relative to a
workspace.

This commit also updates the extraction of configuration scopes as
include paths from the workspace configuration file to replace paths
using these variables.

This allows `$workspace_root` or `$workspace` to refer to the absolute
path to the root of a workspace.