Defer command expansion and fix used variable tracking
This commit defers command expansion until the make_experiments phase,
to allow other phases to modify the variable definitions used in
command expansion (for example, in `mpi_command`).

Additionally, this fixes an issue during the reset of used variables
that caused the dictionary to change while it was being iterated over.