Allow extract paths to be placed in a separate directory
This commit adds a new modifier variable to separate extract paths from
the container path.

The new variable is `container_extract_dir` which is the location
extract paths will be extracted into.

The default is to use `'{workload_input_dir}'` but users can override this,
for more control over where the extracted paths should live.