Fix issue with including command line config scopes
This commit fixes an issue where command line configuration scopes were
not correctly included in resulting workspace configuration scopes.

This was becuase when the workspace module was imported, the
configuration scopes were added (as `ramble.config.get` was called).

Ths changes in this commit delay the `ramble.config.get` call, allowing
the command line scopes to be added.