Migrate lstrip from expander to command suffix
This commit moves an lstrip that was previously generically applied in
the expander as a way to prevent double white space between commands and
their rediction operator.

This commit removes the lstrip from the expander, and adds it on just
the suffix. This allows more consistent formatting in formatted
executables.