Update builtin injection order
This commit updates the order builtins are injected into an executable
graph. Previously, we attempted to prevent cycles by considering
`injection_method` and dependencies separately. However, this makes for
a confusing user experience, and the cycles are ultimately valid based
on how the objects specify the builtins.

This commit updates the logic to have stronger semantics around how
injection_method, depends_on, and dependents are defined.