Allow repeated package definitions with the same name
This commit enables the ability for objects to use software_spec and
define_compiler directives that define the same package name with
different definitions. Previously, this would have given an error, and
now (as seen with Gromacs) repeated specs are allowed as long as they
wouldn't overwrite each other.

Workspace concretize is updated to add prefixes to specs when multiple
package managers are used in the workspace.