Fix defining workflow and package manager variables
Preivously, variables from workflow manager and package managers were
not defined until the add_expand_vars method is called. This means they
are defined quite late in the flow, and sometimes cause issues when
validating experiments.

This commit moves these definitions to when package managers and
workflow managers are added to the experiment.