Allow expander to avoid unescaping escaped braces
This commit adds an option to many of the expansion methods of the
expander class that can be used to define whether escaped curly braces
(i.e. `\{` and `\}`) are unescaped or not during expansion.

This is used in several places to prevent unexpectedly stripping
expansion characters which allows the user to more naturally define the
number of escape characters they are using for their experiments.