Fix workload string representation printing repeated vars
This commit fixes an issue that was introduced with the `when` argument
on variable directives.

After `when` was added, variables were placed in a list within the
workloads (instead of a dict, which they were in before that change).
This change meant that variables were now able to be defined multiple
times within the workload, with the understanding that the last
definition is the one that would actually apply to the workload.

This commit updates this by building a temporary dictionary of variables
in a workload, and using that to print as a string. Additionally, the
variables are now sorted by name instead of being printed by insertion
order.