Add warning capture to ast.parse calls
This commit adds logic to catch warnings and verify they are warning
types we care about when `ast.parse` is called.

Currently, strings that make it into the math evaluation logic with
escaped braces (i.e. '\{' or '\}') cause a SyntaxWarning (that will
later become a SyntaxError). This commit updates this logic so the
warnings are caught and examined before they are printed. Additionally,
a generic `except Execption` block is added to help ensure the future
SyntaxError change doesn't break the expander.