Add front-end command to manage ramble configs
This commit adds the ability for `ramble workspace` to squash all config
sections into a ramble.yaml, and print it to the screen. Additionally,
in the `ramble workspace config` command, features are added to support
simplifying (removing unused) software, and variables.

An issue is fixed in the `get_scope_section` method as well, where
experiment names that contained format specifications would incorrectly
split.