Delay initialize runner for package managers
The motivation comes from `ramble list --type package_managers`
reporting "missing Spack" in an environment where spack is not installed
(for example in doc-build,) even though technically spack is not
required for `ramble list` purpose. Changing it to be a property so that
the "missing in path" error only gets reported upon first-use.