Delay defining experiment variants
This commit updates the logic on defining experiment and default
variants to allow experiment variants to be set even if they do not have
a corresponding default variant. Then when the variant set is
constructed, experiment variants without a default variant are not
included in the resulting set.

This allows sub-objects (such as package managers and workflow managers)
to have their experiment variants handled properly.