Extend to support in memory FOMs
The motiviation is to enable adding FOMs that are not directly backed by
log files. One common use case can be to add such FOMs from the
`_prepare_analysis` hook, without needing to first write them into a log
file.

With the current impl, the inmem FOMs are resolved after the file-based
ones, so in theory this can be used to define derived FOMs down the
line.