Combine variable and variant setting to avoid invalid interactions
This commit updates the base application logic for defining variables
and variants so they happen at the same time. This allows some of their
definition steps to be interleved so they can be used correctly.

Additionally, some changes are made to ensure variants are instance
attributes on the various objects.