Add a tutorial for using `ramble workspace config`
This commit adds a tutorial for the `ramble workspace config` command,
which helps instruct users on how to squash and simplify configuration
files to ease sharing.

Additionally, the squash command is updated to always ignore any of the
repos configuration scopes, as they are not valid in the workspace
configuration section currently.