Update compiler rendering to ensure used variables are tracked
This commit updates the logic around rendering compiler packages.
Previously, the packages were over-aggressively adding variables to the
used variable set, which caused some unused variables to be incorrectly
marked.

This commit updates the logic, and instead causes the rendering of
compilers to happen multiple times to ensure the used variables are
properly tracked.