Simplify env's package rendering
With an environment that contains `m` packages, and a workspace that lists
out `n` packages in total, the previous rendering would call into
`render_package` `O(n * m)` times. The updated flow would instead call `O(n + m)` times. This isn't usually an issue, with `n` and `m` being pretty
small, but can become costly in certain situations. For instance when
the workspace is composed of a long-ish package list.