Refactor initialization of experiment results
This commit refactors the initialization of experiment results.
Previously, the init happened really close to when the results were
populated. This caused some experiments (i.e. children of repeats that
were not executed) to have invalid results objects that were still
accessed (i.e. by their parents). This caused an exception to be thrown
in this (arguably rare) instance.

This refactor avoids this by splitting experiment results into an init
and finalize step.