Update workspace simplify to allow includes
Previously, if an include should have a variable or software
definition removed, the simplify methods would error.

This commit updates the logic in simplify to handle these cases more
gracefully.

Additionally, if simplify would not make any changes, a message is
printed for the users to understand that their config was left
unchanged.