Enhance existing package managers
This commit adds the name of the package manager to software inventory
information, which allows it to be more easily understood when multiple
package managers are used in the same experiment.

Additionally, some python enhancements are made to the spack and pip
package managers. This allows spack to install a python, and tell Ramble
where this python is. The pip package manager is updated to allow
telling it which python to use as well. These changes are to prepare for
a spack-pip package manager, where spack manages the python installation
and pip can then install packages using this installed python.