Add package manager controls for unprefixed specs
This commit updates the pacakge manager base class, and software
environments classes to allow control over whether a package manager
allows unprefixed specs or not.

This is to help support the case where multiple package managers are
used in the same experiment. This can prevent an unprefixed spec from
acceidentally being applied to both package managers.