Use `coverage run` for more accurate coverage numbers
Previously when using `pytest-cov`, it did not capture lines that are executed as part of module initializations. One reason of relying on `pytest-cov` is becuase that handles subprocess calls properly.
In newer versions of `coverage.py`, the "subprocess" patch config now handles this correctly, so we can use `coverage run` directly for a more accurate coverage report.