Add logic for handling conflicts in modifiers
This commit add logic to modifiers, and the modifier language to define
when modifiers conflict with other modifier definitions.

Modifier's will check if they conflict with other existing modifier
definitions when their instances are being created. This will provide
users a better experience when using modifiers that cannot be repeated
on experiments.