Update hashing semantics
This commit updates the hashing semantics to cause all pipelines to
compare a computed hash with any written hash and error if they are
different.

The default exception is the setup pipeline which will create an
inventory.

A global flag is added `--overwrite-inventories` to cause any pipeline
to re-write the inventory / hash. This can also be set in a global
configuration through the config configuration section.